qsort: add benchmarks from the standard library
These will let us be more confident that a custom sort algorithm makes
sense to use, especially when e.g. the Go standard library introduces
a new sorting algorithm, we can compare the two.

On my Mac, using the pdqsort from Go 1.19, the relevant timings are:

    Sort8/100000-10             4.91ms ± 0%
    Sort8/1000000-10            60.1ms ± 0%
    StdlibSort8/100000-10       8.16ms ± 0%
    StdlibSort8/1000000-10       103ms ± 0%

    Sort8/100000-10            163MB/s ± 0%
    Sort8/1000000-10           133MB/s ± 0%
    StdlibSort8/100000-10     98.0MB/s ± 0%
    StdlibSort8/1000000-10    77.5MB/s ± 0%

Compared to the Go 1.18 sort algorithm:

    Sort8/100000-10             4.90ms ± 0%
    Sort8/1000000-10            59.0ms ± 0%
    StdlibSort8/100000-10       8.53ms ± 0%
    StdlibSort8/1000000-10       105ms ± 0%

    Sort8/100000-10            163MB/s ± 0%
    Sort8/1000000-10           136MB/s ± 0%
    StdlibSort8/100000-10     93.8MB/s ± 0%
    StdlibSort8/1000000-10    76.0MB/s ± 0%

So the Go standard library sort performance is improving but we are
still about 40% faster than the benchmark.